+++ src/pages/index.tsx
@@ -1,57 +1,27 @@
-import { useEffect, useState } from "react";
-import CollectionClient from "../backend/db/CollectionClient";
 import Button from "../components/Button";
 import Form from "../components/Form";
 import Layout from "../components/Layout";
 import Table from "../components/Table";
-import Client from "../core/Client";
-import ClientRepository from "../core/ClientRepository";
+import useClients from "../hooks/useClients";
 
 export default function Home() {
 
-  const repo: ClientRepository = new CollectionClient()
-  
-  const [visible, setVisible] = useState<'table' | 'form'>('table') // 2 possiveis estados table ou form
-  
-  const [client, setClient] = useState<Client>(Client.void())
-  
-  const [clients, setClients] = useState<Client[]>([])
-
-  useEffect(getall, [])
-  
-  function getall() {
-    repo.getAll().then((clients) => {
-      setClients(clients)
-      setVisible('table')
-    })
-  }
-  
-  function selectedClient(client: Client) {
-    setClient(client)
-    setVisible('form')
-  }
-  
-  async function deletedClient(client: Client) {
-    await repo.del(client)
-    getall()
-  }
-
-  function newClient() {
-    setClient(Client.void())
-    setVisible('form')
-  }
-
-  async function saveClient (client: Client) {
-    await repo.save(client)
-    getall()
-  }
-
+  const {
+      client,
+      clients,
+      selectedClient,
+      saveClient,
+      deletedClient,
+      newClient,
+      tableVisible,
+      showTable,
+    } = useClients()
   
   return (
     <div className="flex items-center justify-center h-screen bg-gradient-to-r from-sky-600 to-cyan-600">
       
       <Layout titulo="Simple Registration">
-        {visible === 'table' ? (
+        {tableVisible ? (
           <>
             <div className="flex justify-end">
               <Button onClick={() => newClient()}
@@ -61,7 +31,7 @@ export default function Home() {
           </>
         ) : (
           <Form client={client}
-            cancel={() => setVisible('table')}
+            cancel={showTable}
             clientChanged={saveClient}
           />
         )}
